battle
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@222 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/monster/monster.cpp
@@ -649,7 +649,7 @@ Monster::level_t Monster::GetLevel(monster_t monster)
 /* get rnd count monster */
 u16 Monster::GetRNDSize(monster_t monster)
 {
-    u8 level = 0;
+    float level = 0;
 
     switch(Settings::Get().GameDifficulty())
     {
